+++ packages/template/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.0.2
+
+- micromark -> marked (zero dependencies)
+- use module format instead of commonjs
+
 # v1.0.1
 
 - allow HTML in markdown sub-templates